fix(gate): unbreak the file-size ratchet, and retire the entry #1744 should never have added
main is red. `make gate` fails on every branch cut from it:

  crates/stella-core/src/driver.rs grew to 2705 lines, over its baseline
  ceiling of 2689 (+16)
  crates/stella-tools/src/registry.rs grew to 2249 lines, over its baseline
  ceiling of 2214 (+35)

Two things, both about the same guard.

## The ceilings

#1776 added 16 lines to `driver.rs` and 35 to `registry.rs`, both
grandfathered god files. That growth is the case the ratchet documents an
escape hatch for — "raising an existing ceiling is allowed but is never
silent" — so this regenerates the baseline and lands the two-line diff for
review, which is the visibility the guard exists to produce. The feature is
already on main; re-splitting someone else's merged work to save 51 lines is
not the trade the ratchet is asking for.

## The entry that should not exist

`db8cedc4` added a NEW baseline entry, `1534 crates/stella-core/src/loop_detect.rs`.
That file was under the limit before #1744 (mine) landed three witness tests
in its inline `mod tests`, pushing it to 1534. AGENTS.md is explicit that the
baseline accepts no new entries — "a file approaching the limit gets split,
not grown over it" — so the correct response to my PR was a split, and the
entry papered over it instead. That is a defect against my PR, so it is mine
to fix.

`loop_detect.rs`'s tests move to `loop_detect/tests.rs`, the same shape
`driver.rs` / `driver/tests.rs` already uses. The module keeps 558 lines and
its tests 970, so both have real headroom rather than sitting at a ceiling,
and the baseline drops from 33 entries to 32 — the direction it is only ever
supposed to move.

No test content changed: `cargo test -p stella-core --lib loop_detect` reports
the same 48 passing tests before and after the move.

`make guards-fast` is green, including `god-files` (24 god files, named
identically in AGENTS.md and every crate README — `loop_detect.rs` never
reached those tables) and `gate-parity`. `cargo clippy -p stella-core
--all-targets -- -D warnings` and `cargo fmt --check` clean.

Refs #1645, #1744, #1776
