+++ scripts/file-size-baseline.txt
@@ -20,9 +20,8 @@
 4740 crates/stella-cli/src/command_deck.rs
 1507 crates/stella-cli/src/fleet_cmd.rs
 2129 crates/stella-core/src/bus.rs
-2689 crates/stella-core/src/driver.rs
+2705 crates/stella-core/src/driver.rs
 3680 crates/stella-core/src/driver/tests.rs
-1534 crates/stella-core/src/loop_detect.rs
 1780 crates/stella-model/src/anthropic/tests.rs
 2093 crates/stella-model/src/openai.rs
 1565 crates/stella-model/src/zai.rs
@@ -34,7 +33,7 @@
 2268 crates/stella-store/src/tests.rs
 1916 crates/stella-store/src/usage.rs
 1569 crates/stella-tools/src/media.rs
-2214 crates/stella-tools/src/registry.rs
+2249 crates/stella-tools/src/registry.rs
 1839 crates/stella-tools/src/scripts.rs
 1528 crates/stella-tui/src/deck_render.rs
 4006 crates/stella-tui/src/deck_ui.rs
